For my own sanity, rearrange some C++ code.
Probably won't compile yet...
