+++ src/util/helpers.h
@@ -42,24 +42,4 @@ namespace helpers
 
 	//void addCycleKeyword(const std::string& form_spec);
 	//void removeCycleKeyword(const std::string& form_spec);
-
-	struct MenuSelection
-	{
-		static uint32_t getSelectionFromMenu(RE::UI*& a_ui, MenuSelection*& outSelection);
-		static uint32_t makeFromFavoritesMenu(RE::FavoritesMenu* menu, MenuSelection*& outSelection);
-		static void makeFromMagicMenu(RE::MagicMenu* menu, MenuSelection*& outSelection);
-		static void makeFromInventoryMenu(RE::InventoryMenu* menu, MenuSelection*& outSelection);
-
-		MenuSelection(RE::FormID formid);
-		MenuSelection(RE::TESBoundObject* boundObject);
-
-		RE::FormID form_id;
-		bool favorite;
-		bool poisoned;
-		bool equipped;
-		RE::FormType formType;
-		uint32_t count;
-		RE::TESBoundObject* bound_obj;
-		RE::TESForm* form;
-	};
 }
